pyproject.toml
[tool.poetry]
name = "strutex"
version = "1.1.0"
description = "Structured AI document processing with robust fallback strategies."
authors = ["aquiles <achillezongo07@gmail.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
homepage = "https://github.com/Aquilesorei/strutex"
repository = "https://github.com/Aquilesorei/strutex"
keywords = ["pdf", "ai", "llm", "extraction", "json"]

[tool.poetry.plugins."strutex.providers"]
"gemini" = "strutex.providers.gemini:GeminiProvider"
"openai" = "strutex.providers.openai:OpenAIProvider"
"anthropic" = "strutex.providers.anthropic:AnthropicProvider"
"ollama" = "strutex.providers.ollama:OllamaProvider"
"groq" = "strutex.providers.groq:GroqProvider"
"langdock" = "strutex.providers.langdock:LangdockProvider"

[tool.poetry.plugins."strutex.validators"]
"schema" = "strutex.validators.schema:SchemaValidator"
"sum" = "strutex.validators.sum:SumValidator"
"date" = "strutex.validators.date:DateValidator"
"provenance" = "strutex.validators.provenance:ProvenanceValidator"

[tool.poetry.scripts]
strutex = "strutex.cli:main"

[tool.poetry.dependencies]
python = "^3.12"
pypdf = "^6.4.0"
pdfplumber = "^0.11.8"
pandas = "^2.3.3"
openpyxl = "^3.1.5"
pydantic = "^2.12.5"
python-dotenv = "^1.2.1"
pluggy = "^1.5.0"

# --- Optional Dependencies ---
google-genai = { version = "^1.53.0", optional = true }
openai = { version = "^2.8.1", optional = true }
ollama = { version = "^0.6.1", optional = true }
anthropic = { version = "^0.75.0", optional = true }
groq = { version = "^1.0.0", optional = true }
sentence-transformers = { version = "^3.3.0", optional = true }
numpy = { version = "^2.1.0", optional = true }
click = { version = "^8.1.0", optional = true }
pdf2image = { version = "^1.17.0", optional = true }
pytesseract = { version = "^0.3.10", optional = true }
langchain-core = { version = "^1.2.5", optional = true }
llama-index-core = { version = "^0.14.0", optional = true }
haystack-ai = { version = "^2.0.0", optional = true }
unstructured = { version = "^0.11.0", optional = true }
gliner = { version = "^0.2.0", optional = true }
fastapi = {version = "^0.128.0", optional = true}
uvicorn = {version = "^0.40.0", optional = true}
python-multipart = {version = "^0.0.21", optional = true}
fastembed = { version = "^0.7.4", optional = true }
langchain = { version = "^1.2.1", optional = true }
langchain-community = { version = "^0.4.1", optional = true }
langchain-google-genai = { version = "^4.1.3", optional = true }
langchain-openai = { version = "^1.1.6", optional = true }
langgraph = { version = "^1.0.5", optional = true }
qdrant-client = { version = "^1.16.2", optional = true }

[tool.poetry.extras]
cli = ["click"]
ocr = ["pdf2image", "pytesseract"]
google = ["google-genai"]
openai = ["openai"]
anthropic = ["anthropic"]
ollama = ["ollama"]
groq = ["groq"]
langchain = ["langchain-core"]
llamaindex = ["llama-index-core"]
haystack = ["haystack-ai"]
fallback = ["unstructured"]
gliner = ["gliner"]
server = ["fastapi", "uvicorn", "python-multipart"]
guardrails = ["sentence-transformers", "numpy"]
rag = ["fastembed", "langchain", "langchain-community", "langchain-google-genai", "langchain-openai", "langgraph", "qdrant-client", "numpy"]
all = [
    "click", "pdf2image", "pytesseract", "google-genai", "openai", "anthropic", 
    "ollama", "groq", "langchain-core", "llama-index-core", "haystack-ai", 
    "unstructured", "gliner", "fastapi", "uvicorn", "python-multipart", 
    "fastembed", "langchain", "langchain-community", "langchain-google-genai", 
    "langchain-openai", "langgraph", "qdrant-client", "sentence-transformers", "numpy"
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.0"
hypothesis = "^6.92.0"
mypy = "^1.8.0"
pyyaml = "^6.0.0"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.0"
mkdocstrings = {extras = ["python"], version = "^0.27.0"}
mike = "^2.1.0"

[tool.coverage.run]
omit = [
    # CLI
    "strutex/cli.py",
    # Integrations (experimental)
    "strutex/integrations/langchain.py",
    "strutex/integrations/llamaindex.py",
    "strutex/integrations/haystack.py",
    "strutex/integrations/unstructured.py",

    "strutex/providers/anthropic.py",
    "strutex/providers/ollama.py",
    "strutex/providers/groq.py",
    "strutex/providers/chain.py",
    "strutex/providers/hybrid.py",
    "strutex/providers/streaming.py",

    "strutex/schemas/*",
    # Non-code
    "docs/*",
    "examples/*",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"